refactor: ProductProvider, CartProvider에서 컨텍스트 관련 로직만 따로 분리
