+++ src/advanced/contexts/product-context/ProductProvider.tsx
@@ -1,44 +1,19 @@
-import { useCallback, useState, createContext, useMemo } from 'react';
+import { useCallback, useState, useMemo } from 'react';
 
-import { productList as initialProductList } from '../../constants/product';
-import { useCreateCartContext } from '../utils/createContext';
 import { useFlashSale } from './hooks/useFlashSale';
 import { useRecommendProduct } from './hooks/useRecommendProduct';
+import { productContext } from './ProductContext';
+import { productList as initialProductList } from '../../constants/product';
 
 import type { Product } from '../../types/product';
 import type { PropsWithChildren } from 'react';
-
-interface ProductContextType {
-  productList: Product[];
-  increaseQuantity: (id: string) => void;
-  decreaseQuantity: (id: string) => void;
-  resetQuentity: (id: string) => void;
-  addLastSaleItem: (item: Product) => void;
-}
-
-export const productContext = createContext<ProductContextType | undefined>(undefined);
-
-export const useGetProductList = () =>
-  useCreateCartContext(productContext, 'useGetProductList', 'productProvider').productList;
-
-export const useIncreaseQuantity = () =>
-  useCreateCartContext(productContext, 'useIncreaseQuantity', 'productProvider').increaseQuantity;
-
-export const useDecreaseQuantity = () =>
-  useCreateCartContext(productContext, 'useDecreaseQuantity', 'productProvider').decreaseQuantity;
-
-export const useResetQuantity = () =>
-  useCreateCartContext(productContext, 'useResetQuantity', 'productProvider').resetQuentity;
-
-export const useAddLastSaleItem = () =>
-  useCreateCartContext(productContext, 'useAddLastSaleItem', 'cartProvider').addLastSaleItem;
 
 export const ProductProvider = ({ children }: PropsWithChildren) => {
   const [productList, setProductList] = useState<Product[]>(initialProductList);
   const [lastSaleItem, setLastSaleItem] = useState<Product | null>(null);
 
-  useFlashSale(productList, setProductList);
-  useRecommendProduct(productList, setProductList, lastSaleItem);
+  useFlashSale({ productList, setProductList });
+  useRecommendProduct({ productList, setProductList, lastSaleItem });
 
   const calculateQuantity = useCallback(
     (id: string, delta: number) => {
